feat: convert the plugin's metaboxes into backwards compatible only for the legacy editor
